src/infra_ingestao/transformer/requirements.txt
flask==3.1.0
gunicorn==23.0.0
pyiceberg[gcsfs]==0.9.1
pyarrow==19.0.1
gcsfs==2025.2.0
google-cloud-storage==2.19.0
sqlalchemy==2.0.36
pg8000==1.31.2
cloud-sql-python-connector[pg8000]==1.20.0